ulong uint8

+++ src/Layers/xrRenderDX10/dx10SH_Texture.cpp
@@ -215,7 +215,7 @@ void CTexture::ProcessStaging()
     flags.bLoadedAsStaging = FALSE;
 
     //	Check if texture was not copied _before_ it was converted.
-    ULONG RefCnt = pSurface->Release();
+    u32 RefCnt = pSurface->Release();
     pSurface = 0;
 
     VERIFY(!RefCnt);
